--- takwimu/takwimu_ui/src/components/FeaturedData/data.js
@@ -1,12 +1,0 @@
-export default [
-  {
-    title: 'South Africa’s Domestic Consumption',
-    description:
-      'South Africa’s education system faces enormous challenges. Since the end of apartheid, only moderate progress has been made to address the chronic imbalances in the schooling system'
-  },
-  {
-    title: 'South Africa’s National Exports',
-    description:
-      'South Africa’s education system faces enormous challenges. Since the end of apartheid, only moderate progress has been made to address the chronic'
-  }
-];
